Doc Gen Update
I was tired of the website questions, problems, disparities between it and in game, having to edit it and the .tbl files for every property that wasn't a basic ass attribute.

Updated to use .json files from the mod directory.

You can still use the .tbl if you want as I didn't remove the compatibility but if it finds JSON it will ignore TBL and give the errors for missing stuff as if the TBL don't exist.

If any problems arise, I'm sure I'll work them out eventually.